Refactor course slider names and fix minor issues
