+++ src/views/Main/CourseContainer/SliderContainer/ProblemList/ProblemList.style.ts
