+++ src/views/Main/CourseContainer/SliderContainer/TodayProblemContainer/SolvedUserInfo/style.ts
