+++ src/views/Main/CourseContainer/SliderContainer/style.ts
